+++ SampleApp/SampleApp/SampleTableViewController.m
@@ -50,15 +50,15 @@
 
 // Note: This method has been deprecated and replaced by separaterTypeAtIndexPath: Backwards compatibility has been built in for the time being.
 
-- (BOOL)isCellJaggedAtIndexPath:(NSIndexPath *)indexPath
-{
-    if (indexPath.row % 7 == 4 ||indexPath.row % 5 == 2) {
-        return YES;
-    }
-    else {
-        return NO;
-    }
-}
+//- (BOOL)isCellJaggedAtIndexPath:(NSIndexPath *)indexPath
+//{
+//    if (indexPath.row % 7 == 4 ||indexPath.row % 5 == 2) {
+//        return YES;
+//    }
+//    else {
+//        return NO;
+//    }
+//}
 
 - (VENSeparatorType)separatorTypeAtIndexPath:(NSIndexPath *)indexPath
 {
